requirements.txt
fastapi
uvicorn
pydantic
sqlalchemy
psycopg2-binary
boto3
pytest
python-dotenv